+++ packages/app/src/components/shared/Header.tsx
@@ -3,7 +3,7 @@ import Container from '../utils/utils';
 
 function Header() {
   return (
-    <Block styles={{ border: '1px solid #3E3E3E' }} padding="1.5rem 0">
+    <Block styles={{ borderBottom: '1px solid #3E3E3E' }} padding="1.5rem 0">
       <Container>header</Container>
     </Block>
   );
