+++ packages/app/src/pages/_app.tsx
@@ -1,7 +1,6 @@
 import '../../styles/globals.css';
 import type { AppProps } from 'next/app';
 import { BreakpointsProvider, Root, SSRProvider } from '@cube-dev/ui-kit';
-import Layout from '../components/shared/Layout';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
